% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GREGWT.R
\name{alignPop}
\alias{alignPop}
\title{alignPop}
\usage{
alignPop(w, p)
}
\arguments{
\item{w}{new computed weights}

\item{p}{population total}
}
\value{
wo align weights
}
\description{
#' Aligns the survey weights with the total population
}
\details{
#TODO: document alignPop
}
\author{
M. Estebna Munoz H.
}

