% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{alignWithTotals}
\alias{alignWithTotals}
\title{alignWithTotals}
\usage{
alignWithTotals(Tx, breaks, pop_totals, pos, name = "standard",
  verbose = FALSE)
}
\arguments{
\item{Tx,}{un-aligned census data}

\item{pop_totals,}{population totals vector}

\item{verbose}{(optional, default=FALSE) be verbose}
}
\value{
Tx_align, align census data
}
\description{
align the census values to total population, the sum of all categories from
all variables has to be equal to the population total.
}
\author{
M. Estebna Munoz H.
}

