% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{fillMissing}
\alias{fillMissing}
\title{fillMissing}
\usage{
fillMissing(Tx, name_breaks, omit = FALSE)
}
\arguments{
\item{Tx, }{data with missing values (dummy values)}

\item{name_breaks, }{breaks defining the different variables}

\item{omit}{(optional, default=FALSE), omit a specific variable}
}
\value{
Tx_out, data with estimated values
}
\description{
fill the missing in a data set based on the distribution of the variable
}
\author{
M. Estebna Munoz H.
}

