% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{getAreaid}
\alias{getAreaid}
\title{getAreaid}
\usage{
getAreaid(census, census_area_id = FALSE, breaks = FALSE,
  pop_benchmark = FALSE, du_benchmark = FALSE, building_benchmark = FALSE,
  pop_total_col = FALSE, verbose = FALSE)
}
\arguments{
\item{census}{binary census survey data}

\item{breaks}{(optional, default=FALSE) breaks of the categories, needed to
estimate the total population of the areas. If FALSE the function will not
compute the total population for the simulation areas.}

\item{pop_benchmark}{(optional, default=FALSE) define the benchmark to be
used for the computation of the total population, pass as a vector
containing the breaks of the benchmark (e.g. \code{pop_benchmark=c(1,5)}).
If FALSE the function will compute total population as the mean of the all
benchmarks.}

\item{pop_total_col}{(optional, default=FALSE) col containing the population
totals}

\item{verbose}{(optional, default=FALSE) be verbose}

\item{census.area.if}{(optional, default=False) position or name of the area
id in the survey. If False a new id vector will be generated for the survey.}
}
\value{
list(id, pos, pop)
}
\description{
Gets the area id from the survey data or generates a new id for the survey.
This function also calculates the population size of each area.
}
\author{
M. Esteban Munoz H.
}

