% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{getPopulation}
\alias{getPopulation}
\title{getPopulation}
\usage{
getPopulation(Tx, breaks, area_id_pos = FALSE, area_id = FALSE,
  pop_benchmark = FALSE, du_benchmark = FALSE, building_benchmark = FALSE,
  pop_total_col = FALSE, verbose = FALSE)
}
\arguments{
\item{Tx}{census data}

\item{breaks}{positions of the variable breaks in the census}

\item{area_id_pos}{(optional, default=FALSE) position of the area id on the
census. If FALSE the function will try to use \code{area_id} as area codes,
if \code{area_id} is also FALSE the function will issue an error.}

\item{area_id}{(optional, default=FALSE) vector of area codes to be used. If
If FALSE the function will try to use \code{area_id} to identify the column
in the census containing the area codes, if \code{area_id_pos} is also FALSE
the function will issue an error.}

\item{pop_benchmark}{(optional, default=FALSE) define the benchmark to be
used for the computation of the total population, pass as a vector
containing the breaks of the benchmark (e.g. \code{pop_benchmark=c(1,5)}).
If FALSE the function will compute total population as the mean of the all
benchmarks.}

\item{pop_total_col}{(optional, default=FALSE) col containing the population
totals}
}
\value{
total_pop population total for each simulation area
}
\description{
Computes the total population of simulation areas using the constrain
variables.
}
\examples{
data("GREGWT.census")
total_pop <- getPopulation(GREGWT.census, c(11,17), area_id_pos=1)
}
\author{
M. Estebna Munoz H.
}

