% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GREGWT.R
\name{logtocsv}
\alias{logtocsv}
\title{logtocsv}
\usage{
logtocsv(file_name = "GREGWT.log")
}
\arguments{
\item{file_name}{(default = FALSE) alternative file name to process, if
FALSE the function will look for a file name GREGWT.log in the root folder.}
}
\description{
This function will try to read the log output created by function GREGWT
with the output_log variable set to TRUE and generate a csv file on the same
path called GREGWT_log.csv.
}
\author{
M. Esteban Munoz H.
}

