% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{makeDummy}
\alias{makeDummy}
\title{makeDummy}
\usage{
makeDummy(X, verbose = FALSE)
}
\arguments{
\item{X}{data frame of categorical variables.}

\item{verbose}{(optional, default=FALSE) be verbose}
}
\value{
X_out data frame of dummy variables
}
\description{
Make dummy variables out of categorical variables.
}
\examples{
data("GREGWT.survey")
dummy.survey <- makeDummy(GREGWT.survey)
}
\author{
M. Esteban Munoz H.
}

