% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{manualModelMatrix}
\alias{manualModelMatrix}
\title{manualModelMatrix}
\usage{
manualModelMatrix(X, type = "all", use_base = TRUE, verbose = verbose)
}
\arguments{
\item{X}{input matrix}

\item{type}{(optional, default="all"). Maintain all categories or create
reference categories}

\item{use_base}{(optional, default=TRUE). Use the R base function
`model.matrix`}

\item{verbose}{(optional, default=FALSE) be verbose}
}
\value{
X_out. Design matrix
}
\description{
function to manually create a design matrix
}
\author{
M. Esteban Munoz H.
}
