% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{plot.prepareData}
\alias{plot.prepareData}
\title{plot.PrepareData}
\usage{
\method{plot}{prepareData}(data_in, ...)
}
\arguments{
\item{prepareData}{object}
}
\description{
Visualize the correlation between categories of all constrains.
}
\examples{
data("GREGWT.census")
data("GREGWT.survey")

simulation_data <- prepareData(
  GREGWT.census, GREGWT.survey,
  survey_id=FALSE,
  pop_benchmark=c(2,12),
  census_categories=seq(2,24),
  survey_categories=seq(1,3)
)

plot(simulation_data)
}
\author{
M. Esteban Munoz H.
}
