% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{plotPrepareData}
\alias{plotPrepareData}
\title{plotPrepareData}
\usage{
plotPrepareData(data.in, ...)
}
\arguments{
\item{prepareData}{object}
}
\description{
Visualize the correlation between categories of all constrains.
}
\author{
M. Esteban Munoz H.
}

