% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Synthetize.R
\name{Synthetize}
\alias{Synthetize}
\title{Synthetize}
\usage{
Synthetize(x, ...)
}
\arguments{
\item{data_in}{(default = FALSE) an object of class gregwt, compute with the
function gregwt}

\item{X}{(default = FALSE) original survey, must be define if data_in is not
data_in object will overwrite this parameter}

\item{w}{(default = FALSE) output weights used to sample the survey.
data_in object will overwrite this parameter. If data_in = FALSE and w =
FALSE all weight will be set to 1.}

\item{benchmarks}{(default = NULL) use extra benchmarks for the creation of
synthetic individuals.}

\item{pop_size_input}{(default = FALSE) total population size. This
parameter will overwrite the parameter on data_in object.}

\item{method}{(default = c('random', 1)) define the method for creating the
synthetic population. (1) c('random', n) will output an array of size
[pop,x,n] where pop is the population size, x are the number of parameters on
the original survey and n are the number of random samples; (2) c('best', n)
will output an array of size [pop,x], this array is the 'best' fit to
benchmarks; (3) c('bestpop', n) similar to best bust gives more attention
to population totals, important if running an integrated reweight.}

\item{HHsize_mean_fit}{(default = 1.8) mean household size.}

\item{max_iter}{(default = 100) maximal iterations}

\item{fit_tolerance}{(default = 0.1) tolerance to achieve a fit}

\item{random_seed}{(default = 12345) seed for random number generator,
needed to reproduce results.}

\item{output}{(default = FALSE) file to attached to the input survey.}

\item{verbose}{(default = FALSE) be verbose.}
}
\value{
result matrix of synthetic population.
}
\description{
Create a synthetic population based on survey weights.
}
\author{
M. Estebna Munoz H.
}

