% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{toBinary}
\alias{toBinary}
\title{toBinary}
\usage{
toBinary(X, reference_col = FALSE, use_base = TRUE, verbose = FALSE)
}
\arguments{
\item{X}{data matrix containing categorical data}

\item{reference_col}{(optional, default=FALSE) variable defining which
columns to use as reference columns. If FALSE the algorithm will
automatically select which categories to use as reference, if TRUE all
categories will be maintain.}

\item{use_base}{(optional, default=TRUE) use base R function}

\item{verbose}{(optional, default=FALSE) be verbose}
}
\value{
X data matrix containing binary data
}
\description{
Transforms categorical data to binary rows, if the data type of a column is
numerical the function will ignore this column and maintain it as numerical.

The function \code{\link{prepareData}} will use this function to transform
the inputed survey data.
}
\author{
M. Esteban Munoz H.
}

